Moved files around, should be able to test gRPC now
