+++ src/CMakeLists.txt
@@ -1,11 +1,12 @@
 get_filename_component(ristretto_proto "${CMAKE_CURRENT_LIST_DIR}/protos/ristretto.proto" ABSOLUTE)
 get_filename_component(ristretto_proto_path "${ristretto_proto}" PATH)
 
-set(ristretto_proto_srcs "${CMAKE_CURRENT_BINARY_DIR}/ristretto.pb.cc")
-set(ristretto_proto_hdrs "${CMAKE_CURRENT_BINARY_DIR}/ristretto.pb.h")
-set(ristretto_grpc_srcs "${CMAKE_CURRENT_BINARY_DIR}/ristretto.grpc.pb.cc")
-set(ristretto_grpc_hdrs "${CMAKE_CURRENT_BINARY_DIR}/ristretto.grpc.pb.h")
+set(ristretto_proto_srcs "${CMAKE_BINARY_DIR}/src/ristretto.pb.cc")
+set(ristretto_proto_hdrs "${CMAKE_BINARY_DIR}/src/ristretto.pb.h")
+set(ristretto_grpc_srcs "${CMAKE_BINARY_DIR}/src/ristretto.grpc.pb.cc")
+set(ristretto_grpc_hdrs "${CMAKE_BINARY_DIR}/src/ristretto.grpc.pb.h")
 
+# TODO get this command working
 add_custom_command(
     OUTPUT "${ristretto_grpc_srcs}" "${ristretto_grpc_hdrs}"
     COMMMAND protoc
