+++ src/server/KaldiInterface.hpp
@@ -4,6 +4,7 @@
 #include "lat/lattice-functions.h"
 
 namespace kaldi {
+int runDecodeServer(int argc, char* argv[]);
 std::string LatticeToString(const Lattice& lat, const fst::SymbolTable& word_syms);
 std::string GetTimeString(int32 t_beg, int32 t_end, BaseFloat time_unit);
 int32 GetLatticeTimeSpan(const Lattice& lat);
